refactor[#68] : views 레이어 slots 로 네이밍 변경
- 컴포넌트들도 slots 로 네이밍 통일
- nextJs 에선 app router 에서 이미 페이지 혹은 view 단위로 사용하고 있기에, 위젯 이상 페이지 이하의 단위인 slots 로 변경
