+++ blog/src/slots/[article]/ui/index.ts
@@ -1,0 +1,2 @@
+export * from "./GuestArticleSlot";
+export * from "./AdminArticleSlot";
