+++ blog/src/slots/[article]/ui/type.ts
