+++ blog/src/slots/article/ui/TempArticleListSlot.tsx
@@ -7,19 +7,19 @@ import { ArticlePreviewCard } from "@/widgets/article/ui";
 
 import {
   useGetInfiniteArticleList,
-  useGetNumberOfArticles
+  useGetNumberOfTempArticles
 } from "@/entities/article/model";
 
 import { useObserver } from "@/shared/lib";
 import { Grid } from "@/shared/ui/Grid";
 
-const EveryArticleList = ({ numOfArticles }: { numOfArticles: number }) => {
+const ArticleList = ({ numOfArticles }: { numOfArticles: number }) => {
   const {
     data: { pages },
     isFetchingNextPage,
     fetchNextPage,
     hasNextPage
-  } = useGetInfiniteArticleList("published", numOfArticles);
+  } = useGetInfiniteArticleList("draft", numOfArticles);
 
   const observerRef = useObserver(() => fetchNextPage());
 
@@ -29,7 +29,7 @@ const EveryArticleList = ({ numOfArticles }: { numOfArticles: number }) => {
         <Grid>
           {pages.map((article) => (
             <Grid.Item key={article.id}>
-              <Link href={`/article/${article.id}`}>
+              <Link href={`/temp/${article.id}`}>
                 <ArticlePreviewCard {...article} />
               </Link>
             </Grid.Item>
@@ -50,7 +50,7 @@ const EveryArticleList = ({ numOfArticles }: { numOfArticles: number }) => {
           <div ref={observerRef} />
         ) : (
           <div className="flex items-center justify-center py-12 text-gray-400">
-            모든 게시글을 가져왔습니다.
+            모든 임시 저장된 게시글을 가져왔습니다.
           </div>
         )}
       </section>
@@ -58,18 +58,18 @@ const EveryArticleList = ({ numOfArticles }: { numOfArticles: number }) => {
   );
 };
 
-export const EveryArticleListPage = () => {
-  const { data: numOfArticles } = useGetNumberOfArticles();
+export const TempArticleListSlot = () => {
+  const { data: numOfArticles } = useGetNumberOfTempArticles();
 
   return (
     <Suspense fallback={<div>Loading...</div>}>
       <header className="flex justify-center">
         <div className="flex items-center gap-2">
-          <h1 className="text-blue-900">전체 게시글 보기</h1>
+          <h1 className="text-blue-900">임시 저장된 모든 게시글 보기</h1>
           <span className="text-gray-500">({numOfArticles})</span>
         </div>
       </header>
-      <EveryArticleList numOfArticles={numOfArticles} />
+      <ArticleList numOfArticles={numOfArticles} />
     </Suspense>
   );
 };
