+++ blog/src/slots/article/ui/index.ts
@@ -1,0 +1,3 @@
+export * from "./EveryArticleListSlot";
+export * from "./ArticleListBySeriesSlot";
+export * from "./TempArticleListSlot";
