+++ blog/src/slots/main/ui/index.ts
@@ -1,0 +1,3 @@
+export * from "./PopularArticleSlot";
+export * from "./LatestArticleSlot";
+export * from "./SeriesListSlot";
