+++ blog/src/slots/write/lib/index.ts
