+++ blog/src/slots/write/lib/useSeriesSelectToggle.ts
