+++ blog/src/slots/write/model/index.ts
