+++ blog/src/slots/write/ui/items/SeriesSelectToggle.tsx
