--- blog/src/views/[article]/ui/index.ts
@@ -1,2 +1,0 @@
-export * from "./GuestArticlePage";
-export * from "./AdminArticlePage";
