--- blog/src/views/article/ui/index.ts
@@ -1,3 +1,0 @@
-export * from "./EveryArticleListPage";
-export * from "./ArticleListBySeriesPage";
-export * from "./TempArticleListView";
