--- blog/src/views/main/ui/index.ts
@@ -1,3 +1,0 @@
-export * from "./PopularArticleView";
-export * from "./LatestArticleView";
-export * from "./SeriesListView";
